학습 도메인 API 및 문서화, 기능 테스트 완료

+++ src/main/java/psam/portfolio/sunder/english/domain/study/model/response/StudySlicingResponse.java
@@ -28,6 +28,6 @@ public class StudySlicingResponse {
     private String studentName;
     private String schoolName;
     private Integer schoolGrade;
-    private Long correctCount;
-    private Long totalCount;
+    private Integer correctCount;
+    private Integer totalCount;
 }
